Rename FastEvent to GPUStatusFlag and split cuda.hpp into multiple files.
