+++ src/cuda/cuda.cpp
@@ -30,7 +30,6 @@
 #include "Al.hpp"
 #include "aluminum/cuda/cuda.hpp"
 #include "aluminum/mempool.hpp"
-#include "aluminum/cuda/helper_kernels.hpp"
 
 namespace Al {
 namespace internal {
@@ -139,98 +138,6 @@ void replace_internal_streams(std::function<cudaStream_t()> stream_getter) {
 
 bool stream_memory_operations_supported() {
   return stream_mem_ops_supported;
-}
-
-FastEvent::FastEvent() {
-  if (stream_memory_operations_supported()) {
-    sync_event = get_pinned_memory<int32_t>(1);
-    // Initialize to completed to match CUDA event semantics.
-    __atomic_store_n(sync_event, 1, __ATOMIC_SEQ_CST);
-    AL_CHECK_CUDA_DRV(cuMemHostGetDevicePointer(
-                        &sync_event_dev_ptr, sync_event, 0));
-  } else {
-    plain_event = get_cuda_event();
-  }
-}
-
-FastEvent::~FastEvent() {
-  if (stream_memory_operations_supported()) {
-    release_pinned_memory(sync_event);
-  } else {
-    release_cuda_event(plain_event);
-  }
-}
-
-void FastEvent::record(cudaStream_t stream) {
-  if (stream_memory_operations_supported()) {
-    // We cannot use std::atomic because we need the actual address of
-    // the memory.
-#ifndef AL_HAS_ROCM
-    __atomic_store_n(sync_event, 0, __ATOMIC_SEQ_CST);
-    AL_CHECK_CUDA_DRV(cuStreamWriteValue32(
-                        stream, sync_event_dev_ptr, 1,
-                        CU_STREAM_WRITE_VALUE_DEFAULT));
-#else
-    throw_al_exception("A serious error has occurred; should not reach this.");
-#endif
-  } else {
-    AL_CHECK_CUDA(cudaEventRecord(plain_event, stream));
-  }
-}
-
-bool FastEvent::query() {
-  if (stream_memory_operations_supported()) {
-    return __atomic_load_n(sync_event, __ATOMIC_SEQ_CST);
-  } else {
-    cudaError_t r = cudaEventQuery(plain_event);
-    if (r == cudaSuccess) {
-      return true;
-    } else if (r != cudaErrorNotReady) {
-      AL_CHECK_CUDA(r);
-      return false;  // Never reached.
-    } else {
-      return false;
-    }
-  }
-}
-
-GPUWait::GPUWait()
-  : wait_sync(get_pinned_memory<int32_t>(1)),
-    wait_sync_dev_ptr_no_stream_mem_ops(nullptr),
-    wait_sync_dev_ptr(0U)
-{
-  // An atomic here may be overkill.
-  // Can't use std::atomic because we need the actual address.
-  __atomic_store_n(wait_sync, 0, __ATOMIC_SEQ_CST);
-
-  if (stream_memory_operations_supported())
-    AL_CHECK_CUDA_DRV(cuMemHostGetDevicePointer(
-                        &wait_sync_dev_ptr, wait_sync, 0));
-  else
-    AL_CHECK_CUDA(
-      cudaHostGetDevicePointer(
-        reinterpret_cast<void**>(&wait_sync_dev_ptr_no_stream_mem_ops),
-        wait_sync,
-        0));
-}
-
-GPUWait::~GPUWait() {
-  release_pinned_memory(wait_sync);
-}
-
-void GPUWait::wait(cudaStream_t stream) {
-  if (stream_memory_operations_supported())
-#ifdef AL_HAS_ROCM
-    launch_wait_kernel(stream, 1, static_cast<int32_t*>(wait_sync_dev_ptr));
-#else
-    launch_wait_kernel(stream, 1, wait_sync_dev_ptr);
-#endif
-    else
-    launch_wait_kernel(stream, 1, wait_sync_dev_ptr_no_stream_mem_ops);
-}
-
-void GPUWait::signal() {
-  __atomic_store_n(wait_sync, 1, __ATOMIC_SEQ_CST);
 }
 
 }  // namespace cuda
